Jan-Suvidha-Portal/requirements.txt
django==4.2.7
pymongo==4.16.0
dnspython==2.8.0
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.0
python-dotenv==1.0.0
requests==2.31.0
